+++ src/cmd/config_offline.go
@@ -1,0 +1,37 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/git-town/git-town/v7/src/cli"
+	"github.com/spf13/cobra"
+)
+
+func offlineCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "offline [(yes | no)]",
+		Short: "Displays or sets offline mode",
+		Long: `Displays or sets offline mode
+
+Git Town avoids network operations in offline mode.`,
+		Run: func(cmd *cobra.Command, args []string) {
+			if len(args) == 0 {
+				isOffline, err := prodRepo.Config.IsOffline()
+				if err != nil {
+					cli.Exit(err)
+				}
+				cli.Println(cli.FormatBool(isOffline))
+			} else {
+				value, err := cli.ParseBool(args[0])
+				if err != nil {
+					cli.Exit(fmt.Errorf(`invalid argument: %q. Please provide either "yes" or "no".\n`, args[0]))
+				}
+				err = prodRepo.Config.SetOffline(value)
+				if err != nil {
+					cli.Exit(err)
+				}
+			}
+		},
+		Args: cobra.MaximumNArgs(1),
+	}
+}
