+++ src/cmd/root.go
@@ -45,4 +45,5 @@ func init() {
 	RootCmd.PersistentFlags().BoolVar(&debugFlag, "debug", false, "Developer tool to print git commands run under the hood")
 	addAbortCmd(RootCmd)
 	addAppendCmd(RootCmd)
+	RootCmd.AddCommand(configCmd())
 }
